feat: replace component with the one from the library
