+++ package.json
@@ -25,6 +25,7 @@
     "@inubekit/hooks": "^1.2.0",
     "@inubekit/icon": "^2.7.2",
     "@inubekit/link": "^1.5.0",
+    "@inubekit/progressbar": "^1.0.5",
     "@inubekit/select": "^2.4.0",
     "@inubekit/skeleton": "^2.2.0",
     "@inubekit/spinner": "^2.3.0",
