+++ src/components/feedback/ProgressCardWithBarDetermined/styles.ts
@@ -1,9 +1,15 @@
 import styled from "styled-components";
 import { inube } from "@inubekit/foundations";
 import { tokens } from "@design/tokens";
+import { ProgressCardWithBarType } from "./types";
 
 interface IStyledModal {
   $withSmallScreen: boolean;
+}
+
+interface IStyledContainerProgressBar {
+  $appearance: ProgressCardWithBarType;
+  $height: string;
 }
 
 const StyledContainer = styled.div`
@@ -23,4 +29,14 @@ const StyledModal = styled.div<IStyledModal>`
   }
 `;
 
-export { StyledContainer, StyledModal };
+const StyledContainerProgressBar = styled.div<IStyledContainerProgressBar>`
+  width: 100%;
+  background-color: ${({ theme, $appearance }) => {
+    return theme?.progressBar?.[$appearance]?.track.color;
+  }};
+  border: 1px solid ${inube.palette.neutral.N40};
+  border-radius: ${tokens.spacing.s050};
+  height: ${({ $height }) => $height};
+`;
+
+export { StyledContainer, StyledModal, StyledContainerProgressBar };
