--- src/design/feedback/ProgressBar/types.ts
@@ -1,12 +1,0 @@
-export const appearances = [
-  "primary",
-  "success",
-  "warning",
-  "danger",
-  "help",
-  "dark",
-  "gray",
-  "light",
-] as const;
-
-export type ProgressBarAppearanceType = typeof appearances[number];
